review comments

+++ processing/src/main/java/io/druid/query/aggregation/last/DoubleLastAggregatorFactory.java
@@ -149,7 +149,7 @@ public class DoubleLastAggregatorFactory extends NullableAggregatorFactory
             long lastTime = buf.getLong(position);
             if (pair.lhs >= lastTime) {
               buf.putLong(position, pair.lhs);
-              buf.putDouble(position + Long.BYTES, pair.rhs);
+              buf.putDouble(position + Double.BYTES, pair.rhs);
             }
           }
